cache files by default
Former-commit-id: 6c634122c6fa0b96c3c5e204f5b1f1c769bfae8d

+++ learn/skl_utils.py
@@ -17,7 +17,7 @@ if 'IMG_OUTPUT' in os.environ:
 # if system environment has a random seed
 random_state = None
 if 'RANDOM_SEED' in os.environ:
-    random_state = float(os.environ['RANDOM_SEED'])
+    random_state = int(os.environ['RANDOM_SEED'])
 
 def balanced_accuracy(y_true, y_pred):
     """Default scoring function: balanced accuracy.
